# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/attendance": {
            "get": {
                "description": "Get attendance data for all events",
                "tags": [
                    "Attendance"
                ],
                "summary": "Get all attendance records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Attendance"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update attendance record",
                "tags": [
                    "Attendance"
                ],
                "summary": "Update attendance record",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create attendance record",
                "tags": [
                    "Attendance"
                ],
                "summary": "Create attendance record",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete attendance record",
                "tags": [
                    "Attendance"
                ],
                "summary": "Delete attendance record",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/attendance/{id}": {
            "get": {
                "description": "Get data from all Brother records in `Brothers` table",
                "tags": [
                    "Attendance"
                ],
                "summary": "Get all Brothers data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "EventID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Brother"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/events/{eventID}/attendance": {
            "patch": {
                "description": "Update attendance using specific resource endpoint",
                "tags": [
                    "Attendance"
                ],
                "summary": "Update attendance record from eventID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EventID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/events/{eventid}/attendance": {
            "get": {
                "description": "Get event and attendance data by eventID",
                "tags": [
                    "Events"
                ],
                "summary": "Get event and attendance data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "eventid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.GetEventAttendance.EventDataAndAttendance"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new event record",
                "tags": [
                    "Events"
                ],
                "summary": "Create new event record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "eventID",
                        "name": "eventid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Event"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.GetEventAttendance.EventDataAndAttendance": {
            "type": "object",
            "properties": {
                "attendance": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EventAttendance"
                    }
                },
                "eventCategory": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "eventID": {
                    "description": "primary",
                    "type": "integer"
                },
                "eventLocation": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                }
            }
        },
        "models.APIResponse": {
            "description": "JSON response format for all API calls",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Omit if nil"
                },
                "message": {
                    "description": "Omit if nil",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Attendance": {
            "type": "object",
            "properties": {
                "attendanceStatus": {
                    "type": "string"
                },
                "brotherID": {
                    "type": "integer"
                },
                "eventCategory": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "eventID": {
                    "type": "integer"
                },
                "eventLocation": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "rollCall": {
                    "type": "integer"
                }
            }
        },
        "models.Brother": {
            "description": "Brother information",
            "type": "object",
            "required": [
                "firstName",
                "lastName",
                "major",
                "rollCall",
                "status"
            ],
            "properties": {
                "badStanding": {
                    "type": "integer"
                },
                "brotherID": {
                    "description": "Primary Key",
                    "type": "integer"
                },
                "className": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "rollCall": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.Event": {
            "description": "Event information",
            "type": "object",
            "properties": {
                "categoryName": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "eventID": {
                    "description": "primary",
                    "type": "integer"
                },
                "eventLocation": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                }
            }
        },
        "models.EventAttendance": {
            "description": "Event Attendance information of a Brother",
            "type": "object",
            "properties": {
                "attendanceStatus": {
                    "type": "string"
                },
                "brotherID": {
                    "type": "integer"
                },
                "eventID": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "rollCall": {
                    "type": "integer"
                }
            }
        }
    }
}